Updated model of fatree. Rehased code to eliminate warnings. Partially fixed bug by removing GetPriority call from SW allocator in IQ router.
git-svn-id: https://nocs.stanford.edu/cgi-bin/svn.cgi/booksim2.0/trunk@642 ce7aba0d-9f81-40a6-a495-cecae84b0515

+++ creditchannel.cpp
@@ -22,7 +22,7 @@ void CreditChannel::SetLatency( int cycles ) {
 
 void CreditChannel::SendCredit( Credit* credit ) {
 
-  while ( (_queue.size() > _delay) && (_queue.front() == 0) )
+  while ( (_queue.size() > (unsigned int)_delay) && (_queue.front() == 0) )
     _queue.pop( );
 
   _queue.push(credit);
